This changes does several things.
First it turns the project into multiple services. What was just called Petclinic is now
the webapp part and moved under WebApplication.
It adds a EmailService using a prior version of Spring that is vulnerable to log4shell
It adds a Log4Shell Service which serves up the JNDI/LDAP vulnerable classes.
It also runs the other end of the reverse shell.

A docker-compose.yaml file has been created which orchestrates the 3 services. But the webapplication ( petclinic ) part can still be run as it used to be.